Add simple shell script to run clang tidy
Although clang tidy uses compile commands json to find referenced
headers it still cannot be used as input to find the files to analyse.
This script uses jq to constrain checking to the files that are
actually compiled by your cmake configuration, i.e. the files
referenced in compile_commands.json

run-clang-tidy script requires the following programs to be
installed:
 - jq -> parsing the compile commands json
 - parallel -> parallelise execution
 - clang-tidy

Usage example, call from anywhere:
    > ./run-clang-tidy <build-path>
Alternatively supply a specific location if clang-tidy is not on PATH
    > TIDY_PATH=/opt/homebrew/opt/llvm/bin/clang-tidy \
        ./run-clang-tidy <build-path>