Add .git-blame-ignore-revs
'git blame' can be configured to ignore specific commits such as
reformatting the codebase with clang format.

We track the commits we want to be ignored in .git-blame-ignore-revs.

You can use git blame from the cli with:
    git blame --ignore-revs-file .git-blame-ignore-revs <FILE>

Or you configure your git config to always use .git-blame-ignore-revs by
setting:
    git config blame.ignoreRevsFile .git-blame-ignore-revs

Or extending your gitconfig with:

[blame]
    ignoreRevsFile = .git-blame-ignore-revs

Note that .git-blame-ignore-revs is recognized by GitHub.