Fix move ctor/oper for FDB
FDBs move constructor and move assignment operator where not usable
without also including FDBBase.h ad call site.

When declaing move ctor/oper as default in the header both methods are
generated inline. FDB has a unique_ptr member to FDBBase and exposes
only the FDBBase forward declaration. Construction of a uniquer_ptr from
an incomplete type is supported but the type must be complete at the
place of destruction.

If calling code includes FDB.h but not the internal FDBBase.h, the
inline generated move ctor/oper will not know the complete type of
FDBBase and generate a compile error.

The fix is to move the default declaration for both methods to FDB.cpp,
here FDBBase.h is included and thus the complete type is visible.