Enable CI for zarr-fdb-interface
In this PR we introduce a new GH Action workflow to build and test z3fdb
wheels for python 3.11 3.12 3.13. We cannot yet create 3.14 packages
because eccodes (used in testing) is not yet available.

The workflow is not using ECMWF resources and is supposed to be replaced
by an integration into our existing workflows.

Co-authored-by: Kai Kratz <kai.kratz@ecmwf.int>