FDB-508 Prevent marking URIs for deletion if they were previously marked as safe
I wanted to avoid doing this if possible, but it seems once subtocs come into play the question "is this .index file fully masked out" becomes hard to answer. Prior to this commit, any index file referred to by a to-be-deleted subtoc was considered masked and therefore deleted, even if the index file is marked as safe (and not masked) in the parent toc.

I have also update the remote test to use sub tocs.