Use plain cmake for chunked_data_view target
Replaces use of ecbuild_add_library with plain cmake for the
chunked_data_view target. This was done because ecbuild_add_library
automatically creates an install target for each library. However we do
not want to install chunked_data_view, this target is used purely
internal, hence it is compiled into a static library which is used
exclusively in chunked_data_view_bindings.

This fixes a build issue in our native python packages where multio
depends on fdb. The generated instal-targets for fdb reference
chunked_data_view.a which will not be installed, but was still listen in
install-targets by ecbuild.